refactor: move learning_quality_gate into astra-turn-core (270 LOC)
The gate's sole dependency is TurnLearningOutcome from astra_turn_core::contracts,
so turn-core is the natural home. The gate filters low-value signals (trivial queries,
derivable read-only tool chains, ambiguous quality) before they hit the learning pipeline.

- astra-turn-core/src/learning_quality_gate.rs (new)
- runtime/src/pipeline/learning_quality_gate.rs → 5-line re-export stub

Stacks on PRs #124-#127.
